refactor: переместить index.js в директорию pages
*  Перенёс index.js в pages, т.к. это корневой файл страницы. При
   добавлении других страниц их скриптовые файлы размещались бы
   здесь же.
*  В переборе форвычем массива карточек заменил добавление карточки
   в контейнер, для которого ранее уже была создана переменная.
*  Привёл комментарии к общему виду.
