+++ index.html
@@ -133,6 +133,6 @@
     </div>
   </div>
 
-  <script type="module" src="scripts/index.js"></script>
+  <script type="module" src="scripts/pages/index.js"></script>
 </body>
 </html>
